+++ integrations/gcalendar/hub.md
@@ -1,0 +1,51 @@
+# Google Calendar Integration
+
+## Description
+
+The Google Calendar Integration allows you to seamlessly interact with Google Calendar within your Botpress bot. This integration provides various actions to manage calendar events, enhancing the functionality of your bot.
+
+## Actions Supported
+
+The Google Calendar Integration supports the following actions:
+
+- `createEvent`: Create a new event on Google Calendar.
+- `listEvents`: Retrieve a list of upcoming events from Google Calendar.
+- `updateEvent`: Update an existing event on Google Calendar.
+- `deleteEvent`: Delete an event from Google Calendar.
+- `getEventDetails`: Get detailed information about a specific event.
+- `listCalendars`: List all available calendars associated with the authenticated Google account.
+
+## Installation and Configuration
+
+To set up the Google Calendar Integration for your Botpress bot, follow these steps:
+
+1. **Google Cloud Project Setup**:
+   - Create a new project in the [Google Cloud Console](https://console.cloud.google.com/).
+   - Enable the Google Calendar API for your project.
+
+2. **OAuth 2.0 Credentials**:
+   - Create OAuth 2.0 credentials for the project.
+   - Configure authorized redirect URIs to match your Botpress server setup.
+
+3. **Botpress Configuration**:
+   - Install the Google Calendar Integration in your Botpress instance.
+   - Configure the integration using the OAuth 2.0 credentials obtained in step 2.
+
+## Usage
+
+Once the Google Calendar Integration is configured, you can use it to manage calendar events within your Botpress bot. Here are some common use cases:
+
+- Schedule appointments or events on Google Calendar.
+- Retrieve upcoming events and display them to users.
+- Update or delete events based on user requests.
+- Access detailed event information for user inquiries.
+
+The integration provides powerful capabilities to enhance your bot's scheduling and event management functionalities.
+
+## Documentation and Resources
+
+For detailed documentation, API references, and examples, please refer to the integration's README file or the Botpress documentation.
+
+If you encounter any issues or have questions related to the Google Calendar Integration, feel free to reach out to our support team or the Botpress community for assistance.
+
+Start using the Google Calendar Integration to streamline your bot's scheduling and calendar management tasks!
