--- integrations/googleCalendar/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@botpresshub/gcalendar", 
-  "scripts": {
-    "type:check": "tsc --noEmit"
-  },
-  "keywords": [],
-  "private": true,
-  "author": "Goose", 
-  "license": "MIT",
-  "dependencies": {
-    "@botpress/client": "workspace:*",
-    "@botpress/sdk": "workspace:*",
-    "google-auth-library": "^9.0.0",
-    "googleapis": "^126.0.1",
-    "zod": "^3.20.6"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.17",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
-  }
-}
