--- integrations/googleCalendar/src/actions/index.ts
@@ -1,15 +1,0 @@
-import { listEvents } from './list-event';
-import { addEvent } from './create-event';
-import { updateEvent } from './update-event';
-import { deleteEvent } from './delete-event';
-import { listCalendars } from './list-calendars';
-import { getCalendar } from './get-calendar';
-
-export {
-  listEvents,
-  addEvent,
-  updateEvent,
-  deleteEvent,
-  listCalendars,
-  getCalendar,
-};
